Remove *.sh 'local' bashism
Use of the 'local' keyword inside functions in order
to reuse variables without affecting the variable's value
external to the function is not part of the POSIX sh
specification.

In some cases simply removing the 'local' statement is enough.

It other cases it is enough to simply change variable names
to not conflict with callers' names.  Recursive function
invocations, however, must use a subshell.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>