paging: improve paging code
Firstly, do not use mkfifo as it's not available everywhere.

Secondly, properly support the notion GIT_PAGER= and PAGER= to
indicate that no paging should occur.

Thirdly, add support for the LV environment variable.

Fourthly, only set LESS and/or LV environment variables if they
are actually unset, but not if they are set to "" (see the
setup_pager function in Git's pager.c source file).