tg-summary.sh: correct ordering of --rdeps output
By default recurse_deps does a postorder traversal which
results in the deepest leaf being output before its parent.
This makes for an unreadable --rdeps output, so the output
lines were reversed to fix that.  While that does indeed
make the output more readable, the dependencies are then
shown in the reverse order they appear in the .topdeps file
which is not good.

Instead use the new recurse_preorder option to output the
--rdeps information in the correct order without needing to
reverse the order of the output lines.  The output is readable
and in the correct .topdeps order.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>