tg-log: attempt to DWIM with arguments
Add some logic so that in most cases the "--" is not necessary.

And when it is necessary (because paths are being specified)
avoid requiring "--" "--" (unless the path being specified
starts with '-').

This makes use of `tg log` much more intuitive the only
remaining minor nit is that if both options and paths are
being specified for `git log` then "--" is required to
separate them whereas it's sometimes optional for `git log`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>