tg-update: handle a conflict when merging a base
When a conflict occurs while merging a base, the user will be
left on a non-refs/heads/... branch.  After resolving the merge,
committing it and then using `exit`, the following 'tg update'
command will fail because HEAD is not pointing into refs/heads/...
somewhere.

Fix this by always checking out the TopGit branch we're working
on right after the user fixes the merge conflict.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>