tg.sh: make get_tree_w work from non toplevel
While creating a temporary tree, git_tree_w runs `git diff`
which will limit it's output to the current directory.

This will omit changed files if they are not in the current
directory (or one of its children).

Fix this by cd'ing to the top level directory first.  Since
the cd is located in a subshell there are no side effects
except where intended.

This corrects `tg patch -w` output when the current directory
is not at the top level and there are working tree changes
located outside the current directory (and its children).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>