tg-export.sh: call rerere again after shell exit
Since d145a7b8 (Call git-rerere after a failing merge as git-merge
does, topgit-0.8), we've been calling git rerere when we have a
failed merge and are about to spawn a shell for the user to fix
it up.

However, we did not then call git rerere again after shell exit
to record the resolution.  Fix that and call git rerere a second
time after the user has fixed the problem and exited the spawned
shell so the merge resolution will be recorded for future use.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>