tg: exclude objecttype from for-each-ref output
`git for-each-ref` avoids looking up items that are not requested
in the output.  The type of the ref requires an extra lookup but is
just ignored by tg even though it's included in the default output
of `git for-each-ref`.

Specify an explicit format for `git for-each-ref` that excludes the
%(objecttype) so the lookup can be skipped.  This provides a speed
boost of at least 15% when running `git for-each-ref`.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>