tg-update.sh: restore speed
After implementing caching, tg update was reaping a huge benefit
from caching results.  However, if any updates actually took place
the cached results could become inaccurate resulting in an
imcomplete update.

That was then corrected by refusing to use cached results when not
in a "read-only" mode.  However, tg update speed dropped again
when everything was up to date.

Correct this by using the new cache management functions to
explicitly turn caching on and off (which flushes it) where
appropriate so that an update with no changes is just as fast
as it was before while retaining the correctness of the update.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>