tg.sh: ref caching efficiency and correctness
Previously sed was used to select an item from the
ref cache, but the result also required some manipulation.

Unfortunately, this was subject to error if the ref name
contained a '.' and there was another ref name that was
identical except for the the character in the same position
as the dot.  The correct ref value may not have been returned.

Fix this problem while retaining the speedy ref cache by
replacing sed with awk.  The result is more efficient (no
post processing required) and is not subject to inadvertent
regular expression character interpretations.

Although barely measurable, another side benefit of the change
seems to be a very very tiny, almost unmeasurable speed increase.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>