pre-commit.sh: exit on error even with older bash
Running this:

    sh -ec '(false); echo bad'

in older versions of bash produces the output 'bad'
rather than nothing as expected.

Some systems have bash installed as /bin/sh.

A pipeline effectively creates a subshell similarly to ()
when the last item of the pipeline is an internal shell
command and it then has the same problem failing to exit
when `set -e` is in effect and the internal command finishes
with a non-zero exit status.

Work around this bug by explicitly testing the value of $?
for 0 after the subshell exists.  If it is not 0 then test
will exit with a non-zero result and the `set -e` will trigger.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>