tg-create.sh: implement new options and semantics
By default tg-create now runs an editor on .topmsg and
shepherds the initial commit on the new branch so it
only contains the .topdeps and .topmsg changes.

The old behavior can be restored by using --no-edit (or -n).

Just the commit itself can be suppressed with --no-commit
(but the editor will still be run on .topmsg in that case).

As a bonus, if tg create is allowed to shepherd the initial
commit then it will also pre-load a suitable commit message
extracted from .topdeps into MERGE_MSG all ready for the next
commit to save typing.

The .topmsg template now includes a Signed-off-by: line. as well.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>