tg-create.sh: support --topmsg and --topmsg-file options
The two new options (with some obvious aliases) provide a way to
set the .topmsg contents of the new branch while still making the
initial sequestration commit and not running the editor.

The provided .topmsg content will be munged (with a warning to
the user) into a suitable "Subject:" header plus body if needed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>