tg: stop symbolic-ref whining
Add a -q option to git symbolic-ref where needed and replace
some uses with verify_topgit_branch instead as it's more robust
and that's what was being attempted (in many cases poorly) with
use of git symbolic-ref.

Being on a detached HEAD should no longer cause whining when
HEAD is not required to be a TopGit-controlled branch.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>