tg-revert.sh: quiet auto-stash error when no TopGit branches exist
If reverting a tag to recover TopGit branches there may not be any
pre-existing branches.  In that case, the auto stash will fail.

Instead implement a new option and pass it to tg tag to suppress
the error in that case since the point of the auto stash is to aid
recovery of lost refs but if there are no pre-existing TopGit refs
then there's nothing to recover so it's okay to silently do nothing.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>