tg-tag.sh: avoid unnecessary consolidation commit
tg tag will synthesize a consolidation commit if necessary in order
to make sure that all the heads being tagged are referenced by the
tag.  Previously it based the decision on the number of branches
being included in the tag data.  That was incorrect.

Instead correctly base the decision on the number of independent
heads found and tag the sole independent head if only one is found
rather than creating a consolidation commit in that case.

As a shortcut, if only a single branch name is given on the
command line then just tag that branch directly without computing
the independent heads since that branch will necessarily refer
to everything being tagged.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>