tg-create.sh: maybe allow non-clean tree
If a single dependency is listed (or implied) and it's the
same as HEAD (what HEAD points to) then do not require a
clean working tree.

Before sequestration tg create never required a clean working
tree even though it should have when not meeting the above condition.

Allow that behavior again.  Also make sure the initial commit
ONLY commits .topdeps and .topmsg even when other things are staged.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>