tg: always use refs/heads/$name when $name is a TopGit branch
It is possible (though unadvised) to have a branch and a tag
with the same name.  Although `tg create` will not do such a
thing unless forced, there's nothing to prevent a matching tag
being created later without any warning.

When dealing with a $name that is known or required to be
a TopGit branch, always make it explicitly refs/heads/$name in
order to disambiguate it from a possible tag of the same name.

Much of this disambiguation was accomplished in an earlier
commit (all calls to ref_exists) and this completes the job
by fixing up the remaining `git rev-parse` calls.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>