tg-depend.sh: new options and better dirty worktree support
New --no-update and --no-commit options to better control
what `tg depend add` actually does.

The blind commit of everything else staged in the index
was previously corrected but that left no means to combine
another change with the commit other than using
`git commit --amend` after the fact.

The new options remedy that lack by allowing the commit
and/or update to be suppressed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>