tg-tag.sh: exclude non-symbolic refs from TOPGIT REFS
While non-symbolic refs are allowed to affect the parents computation
for the possibly-synthesized consolidation commit, they should not
be included in the TOPGIT REFS section since there's no ref to actually
restore them to.

The warning message claims they will not be included, fix the code
to make that actually reflect reality.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>