tg.sh: output out-of-date remote branch on its own line
Out-of-date remote bases are already output on their own line
when calling branch_needs_update.  Do the same thing for remote
branches themselves rather than using a '%' which may be preempted
by a ':'.

Since '%' is actually a valid branch name, eliminating it as a
special flag avoids a potential ambiguity.

In addition, tg-info will now actually correctly show how many
commits need to be merged in from the remote branch without
a lot of extra work and do so in the same manner it displays
commits that need to be merged in from remote bases.

If the head is out of date with respect to both the remote branch
and the local base, we do output the remote branch first for
the convenience of tg-info that wants to treat out-of-date local
bases as though they'd already been merged into the local branch.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>