tg.sh: prevent dependency loops during recurse_deps
If we have already visited the dependency we are about
to recurse into, complain and skip it.

There should not be any dependency loops.

However, if the checks that prevent creation of such
loops have been bypassed, avoid infinite recursion
by noticing the loop and avoiding it with a warning.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>