bashbug: workaround bash bug set -e exit failure
When /bin/sh is actually bash and it's an older bash, if set -e is
active and a subshell exits with a non-zero status or the last
command in a pipeline is a compound command that exits with a
non-zero status, bash will fail to trigger an exit as required by the
standard when set -e is active in such a circumstance.

Simply adding a 'test $? -eq 0' following the problem code will
successfully trigger an exit in such cases.

However, if the actual value of $? needs to be preserved then calling
a function whose sole statement is "return $?" does the job at slight
loss of obviousness when it's used.

The significant fixes here are for pre-commit.sh and tg-update.sh as
those cases can easily be encountered (and pre-commit.sh has already
had a similar fix made).  The other fixes are more precautionary than
likely to address normally encountered conditions.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>