tg: avoid echo option misinterpretation
On some systems echo may strip and interpret leading options
such as '-n', '-e' and '-E'.

Avoid this by defining a new "echol" function that uses printf
to avoid such misinterpretation and either switch to it or
alter the echo arguments so they cannot be misinterpreted as
an option.

Some echo implementations also by default have "echo -e" in effect.
However, all the "escapes" interpreted in that case start with
a '\' character which is not legal in Git ref names.  Therefore
if we know we have a valid Git ref name and it's the only thing
being interpolated into the string and the string is guaranteed
to start with a non-option then a switch is not necessary.

For example, switching "$dep $parent" is not necessary if both
are known to either be empty or have valid Git refs (or portions
thereof) in them because even if "$parent" is empty and "$dep"
contains "-n", the additional space will prevent $dep from
being treated as an option.

Note that "refs/heads/-n" IS a valid Git ref name.

So, unfortunately, this is necessary.  Unfortunate because on some
systems the antiquated shell does not have "printf" as a built-in
which means that the replacement will be slower on those systems.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>