tg: supply --allow-unrelated-histories option to git merge
Since Git v2.9.0 git merge refuses to operate and merge commits
that do not have at least one common ancestor unless given a new,
special option "--allow-unrelated-histories" to do so.

There is, unfortunately, no magic config option or environment
variable that can be set to say "don't break my scripts".

Instead the version of Git must be tested and the option only
supplied when running Git version 2.9.0 or later.  Fortunately
TopGit already knows the Git version and can easily test it so
the three places where "git merge" is called now supply the
magic option when running Git version 2.9.0 or later.

The Git version is checked in initial_setup and auhopt is set
to the magic option if Git is version 2.9.0 or later and otherwise
the empty string so there isn't even one extra process that need
be run to do this.

In many cases TopGit will not actually need this option, but if
a dependency is added that does not share a common ancestor it
will die without this if Git is running version 2.9.0 or later.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>