tg-annihilate: delay all updates until the end
If one of the tg update calls should fail while annihilating a branch
it could be left only partially annihilated and have to be cleaned
up by hand.

Avoid this by delaying all updates until after annihilation is
complete.  That way, even if updates fail (due to failed merges),
the annihilation has already been completed and does not need
to be cleaned up.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>