tg-revert.sh: detach HEAD before reverting its symref
When reverting the ref that a symbolic-ref HEAD points to
(and it's not a --dry-run), detach HEAD with a message first
(no message with --quiet).

If HEAD is a symbolic-ref to an unborn branch, detach HEAD
to a new empty-tree root commit created with the current
date and time instead.

This allows HEAD's symbolic-ref to be successfully reverted
while leaving the index and working tree untouched (tg
revert NEVER alters the index or working tree) AND making it
possible to simply "git checkout" the newly reverted ref
without needing to add "--force" or suffer from unexpected
attempts to merge new, imaginary diffs between the newly
reverted value of the ref HEAD points to and the index
(which is never affected by tg revert).

It also causes an entry to be added to HEAD's reflog so it's
easy to recover from a detached HEAD and nothing is lost.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>