tg.sh: restore non_annihilated_branches support for args
Since 6466221a2714 (refs: always use fully-qualified refs,
2015-10-10, topgit-0.18), the non_annihilated_branches function lost
the ability to take arguments.

The only client that passes arguments is tg-update.sh and then only
if the "-a"/"--all" option is used and even then the default if
no arguments are given is the same as the behavior of
non_annihilated_branches with no arguments.

So the regression is that if "tg update -a pattern..." was used (where
one or more patterns are actually supplied) then all branches would
be updated as though no patterns were supplied.

Fix this by restoring the ability of non_annihilated_branches to take
arguments with the default if no arguments are given being the same
behavior it had before.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>