tg-tag.sh: add --shell to git for-each-ref and eval
In most cases using an unquoted something that expands to a valid Git
ref name as a command line argument is afe.

Inefficient because it will be subject to pathname expansion and field
splitting (which normally cannot happen unless the value of IFS has
been changed).

Since all of '~', '\', '[', '*', and '?' are disallowed in Git ref
names, one would presume that other than the wasteful check to see if
any of those characters are present to determine whether or not
pathname expansion needs to be performed, the value will remain
unchanged.

Unfortunately on systems aflicted with a "/bin/sh" that is actually
bash or a similarly behaving shell, brace expansion will also take
place and '{', '}' and ',' are all allowed within valid Git ref names.

To avoid this problem, add "--shell" to the "git for-each-ref" command
and eval the resulting command line thus avoiding the issue.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>