tg.sh: remove wc_l function
The wc_l function was added to deal with the peculiarity of *BSD
wc utilities that precede the output count with whitespace.

However, it's much simpler and more efficient just to put the
result inside an arithmetic expansion $(( ... )) expression instead.

Therefore replace the four uses of wc_l with a suitable
$(( $(wc -l <...) )) expression and remove the wc_l function
from tg.sh.  It's important to pass the content to be counted to
the wc utility on standard input otherwise a filename ends up in
the output which the code declines to accomodate.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>