tg: add bin-wrappers/tg to run uninstalled tg
Having to always install TopGit in order to run any kind
of tests on it is awkward.  Instead allow special TG_INST_XXXDIR
environment variables to override the paths built-in by the
make process.

Create a bin-wrappers/tg executable as part of the make process
that runs the freshly built tg with the appropriate environment
variable settings so that the freshly built files are used thus
allowing arbitrary tests on TopGit to be run without installation.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>