Makefile: switch from TG-PREFIX to TG-BUILD-SETTINGS
The TG-PREFIX file was used to track build setting changes
in order to trigger the necessary rebuild when they were modified.

However, it wasn't just the installation "prefix" that was being
tracked and the settings were not being stored in such a way that
they could be reused by any tests (previously there were none).

Therefore replace TG-PREFIX with TG-BUILD-SETTINGS that contains
shell-sourcable variable assignments so it can be reused by tests.

Furthermore, move the definition of GIT_MINIMUM_VERSION from tg.sh
into Makefile so that it too can be tracked and reused by tests.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>