test-lib-main.sh: split initialization into generic and specific
Turn test_lib_main_init into a stub that simply calls first
test_lib_main_init_generic and then test_lib_main_init_specific.

The idea is that anything in test_lib_main_init_generic should NOT be
specific to any particular test when a suite of tests is being run.

For example, parsing the option "--verbose" and setting verbose=t
is the same for all the tests in that one test run.

However, creating the trash directory named after the test and
setting up a tee if requested and redirecting output is test
specific.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>