test-lib-{main,functions}.sh: eliminate use of expr
Avoid spawning several expr processes for tests that
can be accomplished with a parameter expansion and test.

In the general case this isn't possible, but the checks
in question are only looking for the presence of characters
outside a set of allowed characters and that can be done
very easily without using expr.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>