testlib: show failing filename and line number on failures
When a test fails, the script (or whatever else) is output as TAP
diagnostic lines.  Add a line at the beginning that gives the name
of the test script file and the line number of the
"test_expect_success" (or other) command that actually ran the test
that failed.

When multiple jobs are being run and there's test output flying all
over the place, having this nice header line that shows the
location of each failure is invaluable.

Also take this opportunity to clean up the output of the test
script lines a bit by suppressing unwanted blank stuff.  And since
the scripts are being cleaned up when shown also take this
opportunity to do minor cleanup on the help description output too.

Furthermore, add a suitable 'Bail out!' to the end of the error
function that includes file and line number (if available) and
make that 'Bail out!' have the same effect on the
aggregate-results-and-cleanup target that it does when using prove.

Finally, make test_eval_inner_ use (...) instead of {...} so that
any error in the test script doesn't cause premature exit and
instead shows up as a testing error (which will now most likely
include a helpful file and line number to find the problem).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>