testlib: improve lazy prereq cache eval
Make sure that when the lazy prereqs are evaluated at --cache
time that they see the same environment the tests will later
at non-cache time so they generate the correct results.

In particular, it's important to run inside an initialized Git
directory on the "root" (if --root was given) so that the
correct answers for filemode and case insensitive turn up.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>