testlib: find the test library more robustly
The requirement is now that ${TEST_DIRECTORY:-.}/test-lib.sh
always exist and that it will lead (by following a symlink
if necessary) to the correct directory that contains all of
test-lib.sh, test-lib-main.sh and test-lib-functions.sh.

Testing will be aborted if this is not the case.

With this support it's easier to run tests located in other
places by either pointing TEST_DIRECTORY to the right directory
or creating a symlink for test-lib.sh in the current directory.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>