testlib: allow a test script of '-'
If test_expect_success, test_expect_failure or test_tolerate_failure
is passed a test script of '-' then the test script will be read from
standard input.

This allows use of the quoted "HERE" doc format to avoid quoting
nightmares for complicated test scripts.

Unfortunately this is slower -- MUCH slower than passing the test
script directly as a quoted string.  However, the benefit is that
the test script will actually be shown on failure whereas just
putting it in a separate function and passing in the name of the
function will not accomplish that.

Nevertheless, because of the inefficiency (a additional "cat"
process gets spawned), it should be used sparingly and only
when really necessary.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>