testlib: empty tests fail plus improved lint
Make sure that the error code is non-zero when
evaluation of the test script starts so that an
empty test script (which will not change the error
code) always fails.

Also add a "set -e" when running in a subshell to
help exit non-linted tests.

Finally replace the subshell "(exit 117)" with
a much simpler and cheaper "fail_ 117" instead and
add a "${LF}fail_ \$?" suffix so that an empty test
will not fail the test chain lint. (It does not actually
have any broken && chain now does it?)  It will now
just always fail instead as "empty" tests are considered
to be bogus tests since they don't actually test anything.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>