test-lib-functions.sh: make test_must_fail work properly
Now that the test script code normally runs in a subshell (unless
TESTLIB_TEST_NO_SUBSHELL is set) and that subshell tries to activate
"set -e" for the duration, the test_must_fail function requires
updates to work properly.

Firstly most shells ignore "set -e" in this case because they are
buggy and fail to treat "each subshell environment separately." Most
shells leak the "ignore -e while executing the compound list
following while, until, if or elif" incorrectly into the subshell
making the "set -e" wrongly ineffective there.

Secondly bash contains yet another bug (one of many) where using "set
-e" inside an eval string ends up re-activating "set -e" even if it
would have otherwise been inappropriately ignored.  That applies in
this case and so "set -e" is actually active as intended while
running the test code (except when TESTLIB_TEST_NO_SUBSHELL is set)
when using bash (bash's double bugs cancel one another out).

Having "set -e" active while running the "test_must_fail" function
causes it to misbehave as it's not prepared for that and when the
command fails (as it "must"), the "set -e" triggers and the test then
experiences an unexpected fatal abort.

Fix this problem by using a "||" OR list to grab the exit code from
the "must" fail command thereby avoiding any "set -e" triggered exit.

So far so good, except that bash has yet another lovely bug where
using "command" to execute something (as is done here in the git()
function to execute the correct git binary) will result in a "set -e"
triggered exit regardless of whether or not that "command" is inside
a while, until, if or elif condition list or part of an AND/OR list
but not the last command executed.  Yikes!

To fix this problem, change the functions that use "command" (git and
perl) from using '{'...'}' to '('...')' and replace "command" with
"exec".  Then when the failure triggers an exit due to the "set -e"
(even though it shouldn't in this case because it's not the last
command executed in an OR list), the effect will be the same as if it
didn't.

For efficiency the "command" is replaced with "exec" which should
make the result just as efficient as the original '{command ...}'
construct.  Leaving it as "command" should be as efficient as most
shells upon seeing it as the last command in the shell/subshell will
just do an exec automatically, but rather than rely on that
optimization use an "exec" explicitly.  It cannot (and does not need
to) be "exec command" because "command" may only be available as a
shell built-in which would then cause "exec" to complain that it
cannot find "command" and fail (and if it does find it needlessly run
an extra process).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>