t0001: add basic testlib tests
Taken from the Git test suite and adjusted to work
under testlib, t0001 verifies that the testing
framework is operating properly.

It also runs a few tests involving Git which while
not specific to TopGit do help verify that the Git
being used in the tests actually behaves in a sane
fashion.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>