test-lib-main.sh: let --color enable color output
With all TAP output handled by say_tap and say_color_tap
with automatic color removal when `test -n "$HARNESS_ACTIVE"`
there's no reason not to provide a `--color` option to
enable color output to go with the pre-existing `--no-color`
option that disables it.

Therefore provide `--color` as an option.  When HARNESS_ACTIVE
is non-empty color is ALWAYS suppressed for TAP lines so
it's safe to use `--color` with a TAP harness.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>