testlib: never hide error failures
When `test_when_finished` is called from within a subshell it
attempts to diagnose the problem and then call `error` with a
suitable message.

There are two problems with this when running in a subshell:

  1) It's likely a subshell for a test and therefore stdout/stderr
     probably go nowhere
  2) Simply calling `exit` will not exit the testing since it is
     a subshell

Fix #1 by keeping the original stdin, stdout and stderr around
ALWAYS on 6, 5 and 7.  (6 and 5 were already there, 7 is new.)
And make sure that's done right away even before checking the
test_description variable.  Then make sure that the `error`
function writes the error to >&7 and the "Bail out!" to >&5.
While fixing this for the `error` function also make the
`test_debug` function output always go to stderr so that when
it's enabled with `--debug` it can actually be seen.

As part of this fix, we also disable test_when_finished when
doing the test chain lint as if the '&&' chain is broken it
could be called and lead to a bogus and misleading error result
since the linting is always done in a subshell to trap fatal
shell errors.

Fix #2 by adding a trap on SIGUSR1 that always causes a silent
`exit 1` to happen.  Then have the `error` function `kill -USR1 $$`.
Since `$$` will always be the "parent" shell this means the error
and bail messages will be displayed and then the test will exit
one way or another without doing anything else just like it should.

These fixes exposed a minor problem in t0001 that required a tiny
tweak to correct to make t0001 pass again.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>