tg-tag.sh: when using -g on tgstash show "commit" message
The option "--commit-message" is actually a misnomer.  If the
object in question is a commit it does show the commit message,
but if the object is a tag it shows the first line of the tag's
message NOT anything the tag might point to.

When using the tgstash facility the first line of the tag's
message is extremely helpful when needing to "undo" something
whereas the --reflog-message really isn't much help.

Therefore, when viewing the ref log of refs/tgstash default
to using the "--commit-message" option if neither option was
given rather than the normal default of "--reflog-message" for
any other ref log.

There was some code in there that was doing exactly the opposite
but that's just wrong.  Now it does "the right thing"(tm).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>