tg-update.sh: optimize recursion
With a complicated branch structure the update process can
end up spending a lot of time recursing into sub-updates
for dependents.

Speed this up a bit and bypass most of the overhead for those
recursive updates by running them directly from a subshell
rather than spawning a whole new copy of tg with all the
associated startup and setup costs that entails.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>