tg: use --count with rev-list
When the actual value returned by git rev-list is uninteresting,
there are minor efficiencies in the code that can be activated
by passing in --count.  Do so since it's only the number of
items being returned in these cases that are interesting.  In
particular whether there was at least one result.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>