tg-update.sh: merge into base all at once
Instead of merging each out-of-date dependency into the base
immediately after it's been updated, wait until after all the
dependencies have been updated and then merge them into the
base all at once in the same order they would have been if
they'd been merged in one-at-a-time.  This is just begging to
be done as an octopus merge when more than one dependencie is
out-of-date.

This should also eliminate a lot of unnecessary checkouts
and updates of the working tree that can happen in a complex
dependency tree.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>