tg.sh: new --top-bases option to show refs/$topbases
If `tg --top-bases` is run (possibly with one or more
`-c` or `-C` options before the `--top-bases` part), then
show the fully-qualified refs base for top-bases.

This is really just a maintenance command but it is useful
to see what TopGit thinks the "refs/$topbases" value should
be for the current repository.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>