tg-update.sh: reduce number of checkouts slightly
Avoid running checkout (or its equivalent switch_to_base)
unless absolutely necessary.

Add a final checkout to the single branch update mode to
make sure tg-update ends up on the expected branch in that
case.

Do some minor cleanup while in there to the way the
"msg"s are set and passed in to the merge command.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>