tg-update.sh: perform index-only merges whenever possible
Trivial and many simple merges can be done using strictly an index
and, when simple merging is actually required, a couple of
temporary files.

This obviates the need to perform a real checkout.

When a trivial aggressive automatic merge is unable to complete
the merge successfully, simply fallback entirely silently to the
normal working-tree-based merge process (with its associated
performance draining checkout).

This approach substantially reduces the run time for tg update
especially when used on complex dependency trees or when the base
upon which the changes are built has experienced significant
churn, especially when full trees consist of 1000's of files.

As a result of this change, it's not uncommon for tg update to run
on complex dependency trees in 20% or less of the time previously
required to complete the update.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>